+++ src/main/java/day6/Race.java
@@ -1,4 +1,4 @@
 package day6;
 
-public record Race(int record, int distance) {
+public record Race(long record, long distance) {
 }
